Separate out the URI handling
