+++ source/vscode/src/azure/commands.ts
@@ -38,10 +38,12 @@ import {
   getAzurePortalWorkspaceLink,
   getJobFiles,
   getPythonCodeForWorkspace,
+  parseConnectionString,
   queryWorkspaces,
   submitJob,
   uploadBlob,
 } from "./workspaceActions";
+import { UriRouteHandler } from "../uriHandler.js";
 
 const workspacesSecret = `${qsharpExtensionId}.workspaces`;
 let extensionUri: vscode.Uri;
@@ -480,17 +482,44 @@ export async function initAzureWorkspaces(context: vscode.ExtensionContext) {
     ),
   );
 
-  return {
-    /**
-     * Adds a workspace connection programmatically (e.g. from a URI handler).
-     * The workspace is persisted and its refresh cycle is started.
-     */
-    async addWorkspace(workspace: WorkspaceConnection) {
+  /**
+   * URI route handler for the "/connectWorkspace" path.
+   *
+   * Expects a `connectionString` query parameter in the connection string format:
+   *   SubscriptionId=<guid>;ResourceGroupName=<name>;WorkspaceName=<name>;ApiKey=<secret>;QuantumEndpoint=<https://...>
+   *
+   * Shows a confirmation modal before saving the workspace.
+   */
+  const connectWorkspaceUriHandler: UriRouteHandler = async (params) => {
+    const connStr = params.get("connectionString");
+    if (!connStr) {
+      vscode.window.showErrorMessage(
+        "No connection string provided in the workspace URI.",
+      );
+      return;
+    }
+
+    const workspace = parseConnectionString(connStr);
+    if (!workspace) {
+      vscode.window.showErrorMessage(
+        "The workspace URI contained an invalid connection string.",
+      );
+      return;
+    }
+
+    const confirmed = await vscode.window.showInformationMessage(
+      `Add quantum workspace "${workspace.name}" to your connections?`,
+      { modal: true },
+      "Add Workspace",
+    );
+    if (confirmed === "Add Workspace") {
       workspaceTreeProvider.updateWorkspace(workspace);
       await saveWorkspaceList();
       startRefreshCycle(workspaceTreeProvider, workspace);
-    },
+    }
   };
+
+  return { connectWorkspaceUriHandler };
 }
 
 type Buckets = {
